+++ RELEASE-NOTES.md
@@ -18,6 +18,7 @@
 * #107 TableSaver.deleteUnmatchedRows - full sync mode 
 * #108 TableDeleter
 * #109 JdbcConnector: Support for custom data converters
+* #110 Support inferring Series type
 
 ## Release 0.7